`Deleted src/app/auth/callback/page.tsx file`
