+++ src/app/auth/callback/route.ts
@@ -1,0 +1,43 @@
+// app/auth/callback/route.ts
+import { createServerClient } from '@supabase/ssr'
+import { cookies } from 'next/headers'
+import { NextResponse } from 'next/server'
+
+export async function GET(request: Request) {
+  const requestUrl = new URL(request.url)
+  const code = requestUrl.searchParams.get('code')
+
+  if (code) {
+    const supabase = createServerClient(
+      process.env.NEXT_PUBLIC_SUPABASE_URL!,
+      process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+      {
+        cookies: {
+          getAll() {
+            return cookies().getAll()
+          },
+          setAll(cookiesToSet) {
+            cookiesToSet.forEach(({ name, value, options }) => {
+              cookies().set(name, value, options)
+            })
+          }
+        }
+      }
+    )
+    await supabase.auth.exchangeCodeForSession(code)
+  }
+
+  // Recuperamos el rol que guardamos antes del login
+  const cookieStore = cookies()
+  const preferredRole = cookieStore.get('preferred_role')?.value
+
+  // Limpiamos la cookie
+  const response = NextResponse.redirect(
+    preferredRole === 'author'
+      ? `${requestUrl.origin}/dashboard`
+      : `${requestUrl.origin}/biblioteca`
+  )
+  response.cookies.set('preferred_role', '', { maxAge: 0 })
+
+  return response
+}
